requirements.txt
requests
PyGithub
gitpython
pytest
web3
python-dotenv
google-generativeai
ollama
rich
